+++ Tests/Editor/CalculateAssetDependencyTests.cs
@@ -189,7 +189,8 @@ namespace UnityEditor.Build.Pipeline.Tests
         CGPROGRAM
 #pragma surface surf Standard fullforwardshadows
 #pragma target 3.0
-#pragma shader_feature _ TEST_DEFINE
+
+        #pragma shader_feature _ TEST_DEFINE
 
         sampler2D _MainTex;