+++ Tests/Editor/GenerateBundlePackingTests.cs
@@ -1,9 +1,11 @@
+using System;
 using System.Collections.Generic;
 using NUnit.Framework;
 using UnityEditor.Build.Content;
 using UnityEditor.Build.Pipeline.Interfaces;
 using UnityEditor.Build.Pipeline.Tasks;
 using UnityEditor.Build.Utilities;
+using UnityEngine;
 
 namespace UnityEditor.Build.Pipeline.Tests
 {
@@ -36,6 +38,25 @@ namespace UnityEditor.Build.Pipeline.Tests
                 objects.Add(obj);
             }
             return objects;
+        }
+
+        string k_TempAsset = "Assets/temp.prefab";
+        GUID k_TempGuid;
+
+        [OneTimeSetUp]
+        public void OneTimeSetUp()
+        {
+            var root = GameObject.CreatePrimitive(PrimitiveType.Cube);
+            var child = GameObject.CreatePrimitive(PrimitiveType.Cube);
+            child.transform.parent = root.transform;
+            PrefabUtility.SaveAsPrefabAsset(root, k_TempAsset);
+            k_TempGuid = new GUID(AssetDatabase.AssetPathToGUID(k_TempAsset));
+        }
+
+        [OneTimeTearDown]
+        public void OneTimeTearDown()
+        {
+            AssetDatabase.DeleteAsset(k_TempAsset);
         }
 
         [Test]
@@ -141,6 +162,52 @@ namespace UnityEditor.Build.Pipeline.Tests
             var references = new List<ObjectIdentifier>(objects);
             GenerateBundlePacking.FilterReferencesForAsset(dep, assetInBundle, references, new HashSet<ObjectIdentifier>() { objects[0] });
             Assert.AreEqual(0, references.Count);
+        }
+
+#if !UNITY_2019_1_OR_NEWER
+        [Test]
+        public void WhenPrefabContainsDuplicateTypes_GetSortedSceneObjectIdentifiers_DoesNotThorwError()
+        {
+            var includes = new List<ObjectIdentifier>(ContentBuildInterface.GetPlayerObjectIdentifiersInAsset(k_TempGuid, EditorUserBuildSettings.activeBuildTarget));
+            var sorted = GenerateBundleCommands.GetSortedSceneObjectIdentifiers(includes);
+            Assert.AreEqual(includes.Count, sorted.Count);
+        }
+#endif
+
+        [Test]
+        public void BuildCacheUtility_GetSortedUniqueTypesForObjects_ReturnsUniqueAndSortedTypeArray()
+        {
+            var includes = ContentBuildInterface.GetPlayerObjectIdentifiersInAsset(k_TempGuid, EditorUserBuildSettings.activeBuildTarget);
+
+            // Test prefab is created using 2 primitive cubes, one parented to the other, so the includes will in turn contain the sequence 2x:
+            Type[] expectedTypes = new[] { typeof(GameObject), typeof(Transform), typeof(MeshFilter), typeof(MeshRenderer), typeof(BoxCollider) };
+            Array.Sort(expectedTypes, (x, y) => x.AssemblyQualifiedName.CompareTo(y.AssemblyQualifiedName));
+
+            var actualTypes = BuildCacheUtility.GetSortedUniqueTypesForObjects(includes);
+            Assert.AreEqual(expectedTypes.Length * 2, includes.Length);
+            Assert.AreEqual(expectedTypes.Length, actualTypes.Length);
+            CollectionAssert.AreEqual(expectedTypes, actualTypes);
+        }
+
+        [Test]
+        public void BuildCacheUtility_GetMainTypeForObjects_ReturnsUniqueAndSortedTypeArray()
+        {
+            var includes = ContentBuildInterface.GetPlayerObjectIdentifiersInAsset(k_TempGuid, EditorUserBuildSettings.activeBuildTarget);
+
+            // Test prefab is created using 2 primitive cubes, one parented to the other, so the includes will in turn contain the sequence:
+            Type[] expectedTypes = new[] { typeof(GameObject), typeof(Transform), typeof(MeshFilter), typeof(MeshRenderer), typeof(BoxCollider),
+                                           typeof(GameObject), typeof(Transform), typeof(MeshFilter), typeof(MeshRenderer), typeof(BoxCollider) };
+            // One catch, the ordering of the expected types is based on the order of includes which is in turn ordered by the local identifier in file.
+            // Since we are generating the prefab as part of the test, and lfids generation is random, we don't know what order they will be returned in.
+            // So sort both expected types lists and compare exact.
+            Array.Sort(expectedTypes, (x, y) => x.AssemblyQualifiedName.CompareTo(y.AssemblyQualifiedName));
+
+            var actualTypes = BuildCacheUtility.GetMainTypeForObjects(includes);
+            Array.Sort(actualTypes, (x, y) => x.AssemblyQualifiedName.CompareTo(y.AssemblyQualifiedName));
+
+            Assert.AreEqual(expectedTypes.Length, includes.Length);
+            Assert.AreEqual(expectedTypes.Length, actualTypes.Length);
+            CollectionAssert.AreEqual(expectedTypes, actualTypes);
         }
     }
 }